Prototypes and small typo solved. To do : typecasting, list and FD

+++ libft/srcs/malloc/ft_memccpy.c
@@ -6,11 +6,11 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 17:13:17 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/14 12:05:38 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/16 21:58:36 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#inclde "../../includes/libft.h"
+#include "../../includes/libft.h"
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
@@ -19,8 +19,8 @@ void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 	i = -1;
 	while (++i < n)
 	{
-		dest[i] = src[i];
-		if (src[i] == c)
+		(int)dest[i] = (int)src[i];
+		if ((int)src[i] == c)
 			return (&(src[i + 1]));
 	}
 	return (NULL);
